add access checker

+++ libraryapi/src/main/java/com/library/api/libraryapi/services/ReviewService.java
@@ -1,17 +1,22 @@
 package com.library.api.libraryapi.services;
 
 import com.library.api.libraryapi.entities.Review;
+import com.library.api.libraryapi.entities.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.nio.file.AccessDeniedException;
 import java.util.List;
 import java.util.UUID;
 
 import com.library.api.libraryapi.repositories.ReviewRepository;
+import com.library.api.libraryapi.repositories.UserRepository;
 
 @Service
 public class ReviewService {
     
     @Autowired ReviewRepository reviewRepository;
+    @Autowired UserRepository userRepository;
 
     public ReviewService() {
         //Constructor for service
@@ -29,7 +34,9 @@ public class ReviewService {
         return reviewRepository.findByUserId(userId);
     }
 
-    public Review addReview(Review review) {
+    public Review addReview(Review review, String userName) {
+        User user = userRepository.findById(UUID.fromString(userName)).get();
+        review.setUser(user);
         return reviewRepository.save(review);
     }
 
@@ -37,12 +44,19 @@ public class ReviewService {
         return reviewRepository.findById(id).get();
     }
 
-    public void deleteReviewById(UUID id) {
+    public void deleteReviewById(UUID id, String userName) throws AccessDeniedException {
+        var review = getReviewById(id);
+        if (UUID.fromString(userName) != review.getUser().getId()) {
+            throw new AccessDeniedException("");
+        }
         reviewRepository.deleteById(id);
     }
 
-    public Review updateReview(UUID id, Review review) {
+    public Review updateReview(UUID id, Review review, String userName) throws AccessDeniedException {
         review.setId(id);
+        if (UUID.fromString(userName) != review.getUser().getId()) {
+            throw new AccessDeniedException("");
+        }
         return reviewRepository.save(review);
     }
 }
